refactor: convert animated icon components from forwardRef to function components
Replace forwardRef pattern with direct function components and explicit ref props for better type safety and consistency. Update prop types from ComponentProps to ComponentPropsWithoutRef and rename files to use -icon suffix for clarity.
