+++ src/components/animated-icons/chevron-down-icon.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { forwardRef, useImperativeHandle } from "react"
+import { useImperativeHandle } from "react"
 import { motion, useAnimation } from "motion/react"
 
 export type ChevronDownIconHandle = {
@@ -8,56 +8,55 @@ export type ChevronDownIconHandle = {
   stopAnimation: () => void
 }
 
-export type ChevronDownIconProps = React.ComponentProps<"svg"> & {
+export type ChevronDownIconProps = React.ComponentPropsWithoutRef<"svg"> & {
+  ref?: React.Ref<ChevronDownIconHandle>
   duration?: number
 }
 
-const ChevronDownIcon = forwardRef<ChevronDownIconHandle, ChevronDownIconProps>(
-  ({ duration = 0.3, ...props }, ref) => {
-    const controls = useAnimation()
+export function ChevronDownIcon({
+  ref,
+  duration = 0.3,
+  ...props
+}: ChevronDownIconProps) {
+  const controls = useAnimation()
 
-    useImperativeHandle(ref, () => {
-      return {
-        startAnimation: () => controls.start("animate"),
-        stopAnimation: () => controls.start("normal"),
-      }
-    })
+  useImperativeHandle(ref, () => {
+    return {
+      startAnimation: () => controls.start("animate"),
+      stopAnimation: () => controls.start("normal"),
+    }
+  })
 
-    return (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        fill="none"
-        stroke="currentColor"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        aria-hidden
-        {...props}
-      >
-        <motion.path
-          d="M6 9L12 15L18 9"
-          variants={{
-            normal: {
-              d: "M6 9L12 15L18 9",
-            },
-            animate: {
-              d: "M6 15L12 9L18 15",
-            },
-          }}
-          initial="normal"
-          animate={controls}
-          transition={{
-            duration,
-          }}
-        />
-      </svg>
-    )
-  }
-)
-
-ChevronDownIcon.displayName = "ChevronDownIcon"
-
-export { ChevronDownIcon }
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      aria-hidden
+      {...props}
+    >
+      <motion.path
+        d="M6 9L12 15L18 9"
+        variants={{
+          normal: {
+            d: "M6 9L12 15L18 9",
+          },
+          animate: {
+            d: "M6 15L12 9L18 15",
+          },
+        }}
+        initial="normal"
+        animate={controls}
+        transition={{
+          duration,
+        }}
+      />
+    </svg>
+  )
+}
