--- src/components/animated-icons/plus.tsx
@@ -1,92 +1,0 @@
-"use client"
-
-import type { HTMLAttributes } from "react"
-import { forwardRef, useCallback, useImperativeHandle, useRef } from "react"
-import { motion, useAnimation } from "motion/react"
-
-import { cn } from "@/lib/utils"
-
-export interface PlusIconHandle {
-  startAnimation: () => void
-  stopAnimation: () => void
-}
-
-interface PlusIconProps extends HTMLAttributes<HTMLDivElement> {
-  size?: number
-}
-
-const PlusIcon = forwardRef<PlusIconHandle, PlusIconProps>(
-  ({ onMouseEnter, onMouseLeave, className, size = 28, ...props }, ref) => {
-    const controls = useAnimation()
-    const isControlledRef = useRef(false)
-
-    useImperativeHandle(ref, () => {
-      isControlledRef.current = true
-
-      return {
-        startAnimation: () => controls.start("animate"),
-        stopAnimation: () => controls.start("normal"),
-      }
-    })
-
-    const handleMouseEnter = useCallback(
-      (e: React.MouseEvent<HTMLDivElement>) => {
-        if (isControlledRef.current) {
-          onMouseEnter?.(e)
-        } else {
-          controls.start("animate")
-        }
-      },
-      [controls, onMouseEnter]
-    )
-
-    const handleMouseLeave = useCallback(
-      (e: React.MouseEvent<HTMLDivElement>) => {
-        if (isControlledRef.current) {
-          onMouseLeave?.(e)
-        } else {
-          controls.start("normal")
-        }
-      },
-      [controls, onMouseLeave]
-    )
-
-    return (
-      <div
-        className={cn(className)}
-        onMouseEnter={handleMouseEnter}
-        onMouseLeave={handleMouseLeave}
-        {...props}
-      >
-        <motion.svg
-          animate={controls}
-          fill="none"
-          height={size}
-          stroke="currentColor"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth="2"
-          transition={{ type: "spring", stiffness: 100, damping: 15 }}
-          variants={{
-            normal: {
-              rotate: 0,
-            },
-            animate: {
-              rotate: 180,
-            },
-          }}
-          viewBox="0 0 24 24"
-          width={size}
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path d="M5 12h14" />
-          <path d="M12 5v14" />
-        </motion.svg>
-      </div>
-    )
-  }
-)
-
-PlusIcon.displayName = "PlusIcon"
-
-export { PlusIcon }
