+++ src/components/animated-icons/volume-icon.tsx
@@ -1,0 +1,64 @@
+"use client"
+
+import { Fragment, useImperativeHandle, useState } from "react"
+import { AnimatePresence, motion } from "motion/react"
+
+export type VolumeIconHandle = {
+  startAnimation: () => void
+  stopAnimation: () => void
+}
+
+export type VolumeIconProps = React.ComponentPropsWithoutRef<"svg"> & {
+  ref?: React.Ref<VolumeIconHandle>
+}
+
+export function VolumeIcon({ ref, ...props }: VolumeIconProps) {
+  const [isHovered, setIsHovered] = useState(false)
+
+  useImperativeHandle(ref, () => {
+    return {
+      startAnimation: () => setIsHovered(true),
+      stopAnimation: () => setIsHovered(false),
+    }
+  })
+
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 24 24"
+      width="24"
+      height="24"
+      fill="none"
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth="2"
+      {...props}
+    >
+      <path d="M11 4.702a.705.705 0 0 0-1.203-.498L6.413 7.587A1.4 1.4 0 0 1 5.416 8H3a1 1 0 0 0-1 1v6a1 1 0 0 0 1 1h2.416a1.4 1.4 0 0 1 .997.413l3.383 3.384A.705.705 0 0 0 11 19.298z" />
+
+      <AnimatePresence initial={false} mode="wait">
+        {isHovered ? (
+          <Fragment key="volume-icon-active">
+            <motion.path
+              animate={{ opacity: 1, transition: { delay: 0.1 } }}
+              d="M16 9a5 5 0 0 1 0 6"
+              initial={{ opacity: 0 }}
+            />
+            <motion.path
+              animate={{ opacity: 1, transition: { delay: 0.2 } }}
+              d="M19.364 18.364a9 9 0 0 0 0-12.728"
+              initial={{ opacity: 0 }}
+              onAnimationComplete={() => setIsHovered(false)}
+            />
+          </Fragment>
+        ) : (
+          <Fragment key="volume-icon-inactive">
+            <path d="M16 9a5 5 0 0 1 0 6" />
+            <path d="M19.364 18.364a9 9 0 0 0 0-12.728" />
+          </Fragment>
+        )}
+      </AnimatePresence>
+    </svg>
+  )
+}
