+++ src/components/theme-toggle.tsx
@@ -7,8 +7,8 @@ import { META_THEME_COLORS } from "@/config/site"
 import { useClickSound } from "@/hooks/soundcn/use-click-sound"
 import { useMetaColor } from "@/hooks/use-meta-color"
 
-import { MoonIcon } from "./animated-icons/moon"
-import { SunMediumIcon } from "./animated-icons/sun-medium"
+import { MoonIcon } from "./animated-icons/moon-icon"
+import { SunMediumIcon } from "./animated-icons/sun-medium-icon"
 import { Tooltip, TooltipContent, TooltipTrigger } from "./base/ui/tooltip"
 import { Button } from "./ui/button"
 import { Kbd } from "./ui/kbd"
