+++ src/features/portfolio/components/pronounce-my-name.tsx
@@ -5,8 +5,8 @@ import { useHotkeys } from "react-hotkeys-hook"
 
 import { trackEvent } from "@/lib/events"
 import { cn } from "@/lib/utils"
-import type { VolumeIconHandle } from "@/components/animated-icons/volume"
-import { VolumeIcon } from "@/components/animated-icons/volume"
+import type { VolumeIconHandle } from "@/components/animated-icons/volume-icon"
+import { VolumeIcon } from "@/components/animated-icons/volume-icon"
 import { useSound } from "@/registry/hooks/sound/use-sound"
 
 export function PronounceMyName({
